Some cleaning

+++ src/main/java/com/neo/mongocachetest/model/ProductSale.java
@@ -1,8 +1,6 @@
 package com.neo.mongocachetest.model;
 
 import com.neo.mongocachetest.enums.CurrencyCode;
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.Digits;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -14,7 +12,7 @@ public class ProductSale extends BaseDocument{
 
 //    @DecimalMin(value = "0.00", inclusive = true)
 //    @Digits(integer=7, fraction=2)
-    public Double amount;
+    public Float amount;
 
     public CurrencyCode currencyCode;
 }
